랭킹 리스트 수정중 좋아요 누르면 셋 값 해줄것 그리고 미리 불러올 값을 db에 넣어둘것

+++ app/src/main/java/com/koko_plan/server/goodtext/SetMsgToUsers2.java
@@ -23,7 +23,7 @@ import static com.koko_plan.main.MainActivity.photourl;
 
 public class SetMsgToUsers2 {
 
-    private static int getcount;
+    private static long getcount;
 
     public static void send(String touser) {
         new Thread(() -> {
@@ -42,7 +42,7 @@ public class SetMsgToUsers2 {
                         DocumentSnapshot document = task.getResult();
                         if (document != null) {
                             if (document.exists()) {
-                                getcount = (int) document.get("getcount");
+                                getcount = (long) document.get("getcount");
                                 getcount ++;
                             } else {
